# api/openapi.yaml
openapi: 3.0.3
info:
  title: Insurance Claims API
  version: 2.6.0
  description: |
    Contract of the claims settlement platform: policies, claims, supporting
    documents, expertise, payouts, customers and tariffs.

    The document is maintained by hand and is the contract of record for the
    `claims-api-quarkus` service. Neither smallrye-openapi nor any other runtime
    generator is on the classpath, so the file is never overwritten at build time.
  contact:
    name: Claims Platform Team
    email: claims-platform@insurer.example.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api
    description: Local developer instance; the /api prefix comes from @ApplicationPath("/api")
  - url: https://claims.insurer.example.com/api
    description: Production
...
security:
  - bearerAuth: []
paths:
  /v1/policies:
    get:
      operationId: listPolicies
      summary: List insurance policies
      tags: [policies]
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PolicyStatus'
        - name: holderId
          in: query
          required: false
          schema:
            type: string
            maxLength: 40
      responses:
        '200':
          description: Page of policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createPolicy
      summary: Register a new policy in DRAFT status
      tags: [policies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyCreateRequest'
      responses:
        '201':
          description: Policy created
          headers:
            Location:
              description: URI of the created policy
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    options:
      operationId: describePolicyOperations
      summary: Discover the operations supported by the policy resource
      tags: [policies]
      security: []
      responses:
        '200':
          description: Supported operations
          headers:
            Allow:
              description: Allowed HTTP methods
              schema:
                type: string
            X-Policy-Statuses:
              description: Comma separated list of policy statuses
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/policies/{id}:
    parameters:
      - $ref: '#/components/parameters/PolicyIdParam'
    get:
      operationId: getPolicy
      summary: Get a single policy
      tags: [policies]
      responses:
        '200':
          description: Policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      operationId: replacePolicy
      summary: Replace the mutable part of a policy
      tags: [policies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyUpdateRequest'
      responses:
        '200':
          description: Policy updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deletePolicy
      summary: Delete a policy
      tags: [policies]
      responses:
        '204':
          description: Policy deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/reference/currencies:
    get:
      operationId: listCurrencies
      summary: Settlement currencies
      tags: [reference]
      security: []
      responses:
        '200':
          description: Reference items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReferenceItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/reference/claim-types:
    get:
      operationId: listClaimTypes
      summary: Claim categories accepted by the intake form
      tags: [reference]
      security: []
      responses:
        '200':
          description: Reference items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReferenceItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/reference/regions:
    get:
      operationId: listRegions
      summary: Federal districts used for rating
      tags: [reference]
      security: []
      responses:
        '200':
          description: Reference items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReferenceItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/tariffs:
    get:
      operationId: listTariffs
      summary: List rating tariffs
      tags: [tariffs]
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Page of tariffs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TariffPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createTariff
      summary: Create a rating tariff
      tags: [tariffs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TariffCreateRequest'
      responses:
        '201':
          description: Tariff created
          headers:
            Location:
              description: URI of the created tariff
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tariff'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/tariffs/{id}:
    parameters:
      - $ref: '#/components/parameters/TariffIdParam'
    get:
      operationId: getTariff
      summary: Get a single tariff
      tags: [tariffs]
      responses:
        '200':
          description: Tariff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tariff'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      operationId: replaceTariff
      summary: Update the rate and the activity flag of a tariff
      tags: [tariffs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TariffUpdateRequest'
      responses:
        '200':
          description: Tariff updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tariff'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/payouts:
    get:
      operationId: listPayouts
      summary: List settlement payouts
      tags: [payouts]
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PayoutStatus'
        - name: claimId
          in: query
          required: false
          schema:
            type: string
            maxLength: 40
      responses:
        '200':
          description: Page of payouts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayoutPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createPayout
      summary: Draft a payout for a claim
      tags: [payouts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayoutCreateRequest'
      responses:
        '201':
          description: Payout drafted
          headers:
            Location:
              description: URI of the created payout
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payout'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/payouts/{id}:
    parameters:
      - $ref: '#/components/parameters/PayoutIdParam'
    get:
      operationId: getPayout
      summary: Get a single payout
      tags: [payouts]
      responses:
        '200':
          description: Payout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payout'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      operationId: replacePayout
      summary: Replace the amount and the destination of a payout
      tags: [payouts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayoutUpdateRequest'
      responses:
        '200':
          description: Payout updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payout'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/payouts/{id}/approve:
    post:
      operationId: approvePayout
      summary: Approve a drafted payout
      tags: [payouts]
      parameters:
        - $ref: '#/components/parameters/PayoutIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayoutApprovalRequest'
      responses:
        '200':
          description: Payout approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payout'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Operator access token issued by the corporate identity provider
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Static key used by the legacy back office
  parameters:
    OffsetParam:
      name: offset
      in: query
      required: false
      description: Index of the first returned element
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    LimitParam:
      name: limit
      in: query
      required: false
      description: Maximum number of returned elements
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 25
...
    PolicyIdParam:
      name: id
      in: path
      required: true
      description: Policy identifier
      schema:
        type: string
        maxLength: 40
...
    TariffIdParam:
      name: id
      in: path
      required: true
      description: Tariff identifier
      schema:
        type: string
        maxLength: 40
...
    PayoutIdParam:
      name: id
      in: path
      required: true
      description: Payout identifier
      schema:
        type: string
        maxLength: 40
...
  responses:
    BadRequest:
      description: The request payload or parameters failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The caller is authenticated but not allowed to perform the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: The operation conflicts with the current state of the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected server failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
        resource:
          type: string
        timestamp:
          type: string
          format: date-time
        violations:
          type: array
          items:
            type: string
    PageMeta:
      type: object
      required: [offset, limit, total]
      properties:
        offset:
          type: integer
          format: int32
        limit:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    MoneyAmount:
      type: object
      required: [value, currency]
      properties:
        value:
          type: number
          format: double
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: RUB
    ReferenceItem:
      type: object
      required: [code, title]
      properties:
        code:
          type: string
        title:
          type: string
        description:
          type: string
    PolicyStatus:
      type: string
      enum: [DRAFT, ACTIVE, SUSPENDED, EXPIRED, TERMINATED]
    Coverage:
      type: object
      required: [code, title, limit, deductible]
      properties:
        code:
          type: string
          maxLength: 40
        title:
          type: string
          maxLength: 160
        limit:
          $ref: '#/components/schemas/MoneyAmount'
        deductible:
          $ref: '#/components/schemas/MoneyAmount'
    Policy:
      type: object
      required: [id, number, holderId, status, validFrom, validTo, premium, coverages]
      properties:
        id:
          type: string
        number:
          type: string
          pattern: '^POL-[0-9]{10}$'
        holderId:
          type: string
        status:
          $ref: '#/components/schemas/PolicyStatus'
        validFrom:
          type: string
          format: date
        validTo:
          type: string
          format: date
        premium:
          $ref: '#/components/schemas/MoneyAmount'
        coverages:
          type: array
          description: Nested collection of coverage objects attached to the policy
          items:
            $ref: '#/components/schemas/Coverage'
        updatedAt:
          type: string
          format: date-time
    PolicyCreateRequest:
      type: object
      required: [number, holderId, validFrom, validTo, premium, coverages]
      properties:
        number:
          type: string
          pattern: '^POL-[0-9]{10}$'
        holderId:
          type: string
          maxLength: 40
        validFrom:
          type: string
          format: date
        validTo:
          type: string
          format: date
        premium:
          $ref: '#/components/schemas/MoneyAmount'
        coverages:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: '#/components/schemas/Coverage'
    PolicyUpdateRequest:
      type: object
      required: [status, validTo, premium]
      properties:
        status:
          $ref: '#/components/schemas/PolicyStatus'
        validTo:
          type: string
          format: date
        premium:
          $ref: '#/components/schemas/MoneyAmount'
        coverages:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/Coverage'
    PolicyPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Policy'
    ClaimType:
      type: string
      enum: [PROPERTY_DAMAGE, THEFT, FLOOD, FIRE, MEDICAL, LIABILITY]
...
    Tariff:
      type: object
      required: [id, code, title, appliesTo, baseRate, minimumPremium, effectiveFrom]
      properties:
        id:
          type: string
        code:
          type: string
          pattern: '^TRF-[A-Z0-9]{4,10}$'
        title:
          type: string
          maxLength: 160
        appliesTo:
          $ref: '#/components/schemas/ClaimType'
        baseRate:
          type: number
          format: double
          minimum: 0.001
          maximum: 1.0
        minimumPremium:
          $ref: '#/components/schemas/MoneyAmount'
        effectiveFrom:
          type: string
          format: date
        active:
          type: boolean
    TariffCreateRequest:
      type: object
      required: [code, title, appliesTo, baseRate, minimumPremium, effectiveFrom]
      properties:
        code:
          type: string
          pattern: '^TRF-[A-Z0-9]{4,10}$'
        title:
          type: string
          maxLength: 160
        appliesTo:
          $ref: '#/components/schemas/ClaimType'
        baseRate:
          type: number
          format: double
          minimum: 0.001
          maximum: 1.0
        minimumPremium:
          $ref: '#/components/schemas/MoneyAmount'
        effectiveFrom:
          type: string
          format: date
    TariffUpdateRequest:
      type: object
      required: [baseRate, minimumPremium, active]
      properties:
        baseRate:
          type: number
          format: double
          minimum: 0.001
          maximum: 1.0
        minimumPremium:
          $ref: '#/components/schemas/MoneyAmount'
        active:
          type: boolean
    TariffPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Tariff'
...
    PayoutStatus:
      type: string
      enum: [DRAFT, PENDING_APPROVAL, APPROVED, SENT, SETTLED, FAILED]
    PayoutDestination:
      type: object
      description: Polymorphic settlement destination, discriminated by the `channel` property.
      required: [channel]
      properties:
        channel:
          type: string
          enum: [BANK_TRANSFER, CARD]
      discriminator:
        propertyName: channel
        mapping:
          BANK_TRANSFER: '#/components/schemas/BankTransferDestination'
          CARD: '#/components/schemas/CardDestination'
      oneOf:
        - $ref: '#/components/schemas/BankTransferDestination'
        - $ref: '#/components/schemas/CardDestination'
    BankTransferDestination:
      type: object
      required: [channel, accountNumber, bankCode, beneficiaryName]
      properties:
        channel:
          type: string
          enum: [BANK_TRANSFER]
        accountNumber:
          type: string
          pattern: '^[0-9]{20}$'
        bankCode:
          type: string
          pattern: '^[0-9]{9}$'
        beneficiaryName:
          type: string
          maxLength: 200
    CardDestination:
      type: object
      required: [channel, last4, scheme, issuerCountry]
      properties:
        channel:
          type: string
          enum: [CARD]
        last4:
          type: string
          pattern: '^[0-9]{4}$'
        scheme:
          type: string
          enum: [VISA, MASTERCARD, MIR]
        issuerCountry:
          type: string
          pattern: '^[A-Z]{2}$'
    Payout:
      type: object
      required: [id, claimId, status, amount, destination]
      properties:
        id:
          type: string
        claimId:
          type: string
        status:
          $ref: '#/components/schemas/PayoutStatus'
        amount:
          $ref: '#/components/schemas/MoneyAmount'
        destination:
          $ref: '#/components/schemas/PayoutDestination'
        approvedBy:
          type: string
          maxLength: 120
        externalReference:
          type: string
        createdAt:
          type: string
          format: date-time
    PayoutCreateRequest:
      type: object
      required: [claimId, amount, destination]
      properties:
        claimId:
          type: string
          maxLength: 40
        amount:
          $ref: '#/components/schemas/MoneyAmount'
        destination:
          $ref: '#/components/schemas/PayoutDestination'
    PayoutUpdateRequest:
      type: object
      required: [amount, destination]
      properties:
        amount:
          $ref: '#/components/schemas/MoneyAmount'
        destination:
          $ref: '#/components/schemas/PayoutDestination'
    PayoutApprovalRequest:
      type: object
      required: [approver]
      properties:
        approver:
          type: string
          maxLength: 120
        comment:
          type: string
          maxLength: 500
    PayoutPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Payout'
...